Добавил поддержку CancellationToken

+++ PipeDiagnosticsViewer.Test/PipeDiagnosticFromFileServiceFixture.cs
@@ -2,6 +2,7 @@
 using PipeDiagnosticsViewer.Infrastructure;
 using PipeDiagnosticsViewer.Models;
 using NUnit.Framework.Legacy;
+using PipeDiagnosticsViewer.Interfaces;
 
 namespace PipeDiagnosticsViewer.Test
 {
@@ -20,8 +21,8 @@ namespace PipeDiagnosticsViewer.Test
         public async Task GetItemsFromCSVTestAsync()
         {
             string filePath = TestHelper.GetTempFilePath("PipeDiagnosticsViewer.Test.Data.objects.csv");
-            List<PipeDiagnostic> pipeDiagnostics = await GetPipeDiagnosticsA(filePath);
-            ClassicAssert.AreEqual(pipeDiagnostics.Count(), 9);
+            List<PipeDiagnostic> pipeDiagnostics = await GetPipeDiagnosticsAsync(filePath, CancellationToken.None);
+            ClassicAssert.AreEqual(pipeDiagnostics.Count, 9);
         }
 
         [Test]
@@ -33,31 +34,60 @@ namespace PipeDiagnosticsViewer.Test
         }
 
         [Test]
-        public async Task GetItemsFromExcelTTestAsync()
+        public async Task GetItemsFromExcelTestAsync()
         {
             string filePath = TestHelper.GetTempFilePath("PipeDiagnosticsViewer.Test.Data.Objects.xlsx");
-            List<PipeDiagnostic> pipeDiagnostics = await GetPipeDiagnosticsA(filePath);
+            List<PipeDiagnostic> pipeDiagnostics = await GetPipeDiagnosticsAsync(filePath, CancellationToken.None);
             ClassicAssert.AreEqual(pipeDiagnostics.Count, 4);
+        }
+
+        [Test]
+        public void MuitiItemsFromCSVTestAsync()
+        {
+            string filePath = TestHelper.GetTempFilePath("PipeDiagnosticsViewer.Test.Data.objects.csv");
+            CancellationTokenSource cancelTokenSource = new CancellationTokenSource();
+            List<PipeDiagnostic> pipeDiagnostics = new List<PipeDiagnostic>();
+            Task task1 = AddPipeDiagnostic(pipeDiagnostics, filePath, cancelTokenSource.Token);
+            cancelTokenSource.Cancel();
+            pipeDiagnostics.Clear();
+            cancelTokenSource = new CancellationTokenSource();
+            Task task2 = AddPipeDiagnostic(pipeDiagnostics, filePath, cancelTokenSource.Token);
+            cancelTokenSource.Cancel();
+            pipeDiagnostics.Clear();
+            cancelTokenSource = new CancellationTokenSource();
+            Task task3 = AddPipeDiagnostic(pipeDiagnostics, filePath, cancelTokenSource.Token);
+            cancelTokenSource.Cancel();
+            pipeDiagnostics.Clear();
+            cancelTokenSource = new CancellationTokenSource();
+            Task task4 = AddPipeDiagnostic(pipeDiagnostics, filePath, cancelTokenSource.Token);
+            Task.WaitAll(task1, task2, task3, task4);
+            ClassicAssert.AreEqual(pipeDiagnostics.Count, 9);
         }
 
         private IEnumerable<PipeDiagnostic> GetPipeDiagnostics(string filePath)
         {
-            PipeDiagnosticFromFileService pipeDiagnosticFromFileService = new PipeDiagnosticFromFileService();
+            IPipeDiagnosticFromFileService pipeDiagnosticFromFileService = new PipeDiagnosticFromFileService();
             IEnumerable<PipeDiagnostic> pipeDiagnostics = pipeDiagnosticFromFileService.GetItems(filePath);
             File.Delete(filePath);
             return pipeDiagnostics;
         }
 
-        private async Task<List<PipeDiagnostic>> GetPipeDiagnosticsA(string filePath)
+        private async Task<List<PipeDiagnostic>> GetPipeDiagnosticsAsync(string filePath, CancellationToken cancellationToken, bool isDeletefile = true)
         {
-            PipeDiagnosticFromFileService pipeDiagnosticFromFileService = new PipeDiagnosticFromFileService();
+            IPipeDiagnosticFromFileService pipeDiagnosticFromFileService = new PipeDiagnosticFromFileService();
             List<PipeDiagnostic> pipeDiagnostics = new List<PipeDiagnostic>();
-             await foreach (PipeDiagnostic pipeDiagnostic in pipeDiagnosticFromFileService.GetItemsAsync(filePath))
+             await foreach (PipeDiagnostic pipeDiagnostic in pipeDiagnosticFromFileService.GetItemsAsync(filePath, cancellationToken))
                 pipeDiagnostics.Add(pipeDiagnostic);
-
-            File.Delete(filePath);
+            
+             if(isDeletefile)
+                File.Delete(filePath);
 
             return pipeDiagnostics;
+        }
+
+        private async Task AddPipeDiagnostic(List<PipeDiagnostic> pipeDiagnostics, string filePath, CancellationToken cancellationToken)
+        {
+            pipeDiagnostics.AddRange(await GetPipeDiagnosticsAsync(filePath, cancellationToken, false));
         }
     }
 }
